+++ packages/xchain-ethereum/__mocks__/index.ts
@@ -1,7 +1,14 @@
 import { mock_etherscan_api } from './etherscan-api'
 import { mock_infra_api } from './infra-api'
 
-export const mock_all_api = (etherscanUrl: string, infraUrl: string, alchemyUrl: string, method: string, result: string | Object) => {
+export const mock_all_api = (
+  etherscanUrl: string,
+  infraUrl: string,
+  alchemyUrl: string,
+  method: string,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  result: any,
+) => {
   mock_etherscan_api(etherscanUrl, method, result)
   mock_infra_api(infraUrl, method, result)
   mock_infra_api(alchemyUrl, method, result)