+++ packages/xchain-ethereum/CHANGELOG.md
@@ -1,5 +1,16 @@
 # v.x.x.x
 
+# v.0.14.2 (2021-03-09)
+
+### Fix
+
+- export Client.`estimateGasPrices`, Util.`getFee`
+- save `etherscanApiKey` in Client
+- update `getTokenAddress` to return checksum address
+
+### Breaking change
+- change parameter of `getTokenAddress`
+
 # v.0.14.1 (2021-03-08)
 
 ### Fix